requirements.txt
attrs==25.4.0
Django==6.0.3
django-filter==25.2
django-stubs==6.0.1
django-stubs-ext==6.0.1
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
drf-spectacular==0.29.0
psycopg2-binary==2.9.11
flake8==7.3.0
flake8-isort==7.0.0
isort==8.0.1
pytest==9.0.2
pytest-django==4.12.0
factory-boy==3.3.3
